mcp_productivity/requirements.txt
# MCP Productivity Server Dependencies
# OpenProject
requests>=2.31.0

# Database
sqlalchemy>=2.0.0

# MCP Server
mcp>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
python-dateutil>=2.8.2
pydantic>=2.5.0
